+++ src/native/flex-native-ali/index.js
@@ -1,5 +1,7 @@
 /* eslint-disable */
 // eslint-disable-next-line no-undef
+import plainStyle from '@remax/one/esm/useWebPlaceholderStyle/plainStyle';
+
 Component({
   props: {
     id: String,
@@ -7,6 +9,18 @@ Component({
     className: String,
     content: String,
     visible: Boolean,
+  },
+  methods: {
+    onTap(e) {
+      this.triggerEvent('tap', e);
+      this.props.onThis({
+        setData: ({ style, ...data }) =>
+          this.setData({
+            ...data,
+            style: typeof style === 'object' ? plainStyle(style) : style,
+          }),
+      });
+    },
   },
   // onInit() {
   //   this.triggerEvent('this', this);