+++ src/use-view-layout/index.tsx
@@ -1,21 +1,24 @@
-import { useSize } from 'ahooks';
 import { useForceUpdate } from 'parsec-hooks';
 import { useEffect, useRef } from 'react';
 
-export const getWH = (id: string) =>
-  new Promise<{ width: number; height: number }>(resolve => {
-    const { offsetWidth, offsetHeight } = document.getElementById(id) || {};
-    resolve({
-      width: offsetWidth || 0,
-      height: offsetHeight || 0,
-    });
-  });
+export const getLayout = (id: string) =>
+  new Promise<{ width?: number; height?: number; x?: number; y?: number }>(
+    resolve => {
+      const { width, height, x, y } =
+        document.getElementById(id)?.getBoundingClientRect() || {};
+      resolve({
+        width,
+        height,
+        x,
+        y,
+      });
+    },
+  );
 
 export default ({ run = true }: { run?: boolean } = {}) => {
   const { forceUpdate } = useForceUpdate();
   const ref = useRef<HTMLElement>(null);
-  const { width, height } = useSize(run ? ref.current : null) || {};
-  const { y, x } = ref.current?.getBoundingClientRect() || {};
+  const { y, width, x, height } = ref.current?.getBoundingClientRect() || {};
   useEffect(() => {
     if (height === undefined && run) {
       forceUpdate();